#1 - Core functionalities: Download all MP3 from YouTube URL file
On startup, the program creates the "YTmp3 Fast Converter" folder in the user's Downloads directory.
Then, the "videos_to_download_urls.txt" file is created, to store the YouTube URLs to download, as well as the "Downloaded music" folder, to store the MP3 files once downloaded.
yt-dlp is used for getting video information and downloading the MP3.
When downloading, the artist and the track information on a video is looked up. If it is found, the file name will follow the pattern "<artist_name> - <track_name>", otherwise it will be the video title.

Close #1