#4 - Fix linter suggestions & Add type hints
Fix incorrectly formatted names for variables, functions and classes.
Add type hints for all the files, with strict mode on.
A stub for yt-dlp (in "typings/yt_dlp") has subsequently been created, in which only used methods are defined.
Add .vscode/ to the .gitignore to avoid spreading IDE-specific settings.

Close #4